@@ -12,28 +12,27 @@
     "lint": "ng lint && stylelint 'src/**/*.scss'",
     "format": "prettier src/**/*.{js,jsx,ts,tsx,css,scss,md} --write",
     "e2e": "ng e2e",
-    "precommit": "lint-staged",
     "bundle-report": "webpack-bundle-analyzer dist/stats.json"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^5.2.9",
-    "@angular/cdk": "^5.2.5",
-    "@angular/common": "^5.2.9",
-    "@angular/compiler": "^5.2.9",
-    "@angular/core": "^5.2.9",
-    "@angular/flex-layout": "^5.0.0-beta.14",
-    "@angular/forms": "^5.2.9",
-    "@angular/http": "^5.2.9",
-    "@angular/material": "^5.2.5",
-    "@angular/platform-browser": "^5.2.9",
-    "@angular/platform-browser-dynamic": "^5.2.9",
-    "@angular/router": "^5.2.9",
+    "@angular/animations": "6.0.0",
+    "@angular/cdk": "6.0.1",
+    "@angular/common": "6.0.0",
+    "@angular/compiler": "6.0.0",
+    "@angular/core": "6.0.0",
+    "@angular/flex-layout": "6.0.0-beta.15",
+    "@angular/forms": "6.0.0",
+    "@angular/http": "6.0.0",
+    "@angular/material": "6.0.1",
+    "@angular/platform-browser": "6.0.0",
+    "@angular/platform-browser-dynamic": "6.0.0",
+    "@angular/router": "6.0.0",
     "@ecodev/fab-speed-dial": "^0.0.4",
-    "@mat-datetimepicker/core": "^1.0.2",
-    "@swimlane/ngx-charts": "^7.2.0",
+    "@mat-datetimepicker/core": "^1.0.4",
+    "@swimlane/ngx-charts": "^8.0.0",
     "angular-calendar": "^0.23.7",
-    "apollo-angular": "^1.0.1",
+    "apollo-angular": "^1.1.0-rc.0",
     "apollo-angular-link-http": "^1.0.2-beta.0",
     "apollo-cache-inmemory": "^1.1.11",
     "apollo-client": "^2.2.7",
@@ -48,16 +47,18 @@
     "jest": "^22.4.2",
     "jest-preset-angular": "^5.2.1",
     "jspdf": "^1.3.5",
-    "material-community-components": "^2.7.8",
-    "rxjs": "^5.5.7",
+    "material-community-components": "^3.0.0",
+    "rxjs": "^6.1.0",
+    "rxjs-compat": "^6.1.0",
     "stylelint": "^9.2.0",
     "webpack-bundle-analyzer": "^2.11.1",
-    "zone.js": "^0.8.19"
+    "zone.js": "^0.8.26"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "^5.2.9",
-    "@angular/language-service": "^5.2.9",
+    "@angular-devkit/build-angular": "~0.6.0",
+    "@angular/cli": "6.0.0",
+    "@angular/compiler-cli": "6.0.0",
+    "@angular/language-service": "6.0.0",
     "@types/html2canvas": "^0.0.33",
     "@types/jest": "^22.2.0",
     "@types/jspdf": "^1.1.31",
@@ -72,7 +73,7 @@
     "ts-node": "~4.1.0",
     "tslint": "~5.9.1",
     "tslint-config-prettier": "^1.10.0",
-    "typescript": "~2.5.3"
+    "typescript": "2.7.2"
   },
   "majestic": {
     "jestConfig": "./jest.config.js"
